rtl_fm: Add squelch timeout
The current squelch logic waits until the received signal drops below
the squelch value and then continues the frequency scan. If there's any
continuous transmission on the air, it will block the scanning process
forever. This patch adds support for squelch timeout to prevent the
scanner from locking forever to a single signal.

Signed-off-by: Nikolay Dimitrov <picmaster@mail.bg>